[REFACTOR/#222] feature:navigator 모듈 삭제 및 app 모듈에 권한 위임
