+++ app/build.gradle.kts
@@ -29,21 +29,24 @@ android {
 
 dependencies {
     implementation(projects.core.common)
+    implementation(projects.core.analytics)
     implementation(projects.core.buildconfig)
     implementation(projects.core.network)
     implementation(projects.core.designsystem)
     implementation(projects.core.datastore)
     implementation(projects.core.alarm)
     implementation(projects.core.media)
+    implementation(projects.core.ui)
     implementation(projects.data)
     implementation(projects.domain)
+    implementation(projects.feature.splash)
     implementation(projects.feature.onboarding)
     implementation(projects.feature.home)
     implementation(projects.feature.alarmInteraction)
     implementation(projects.feature.fortune)
     implementation(projects.feature.mission)
     implementation(projects.feature.setting)
-    implementation(projects.feature.navigator)
+    implementation(projects.feature.webview)
     implementation(libs.firebase.analytics)
     implementation(libs.firebase.crashlytics)
     implementation(libs.play.services.ads)
