+++ app/src/main/java/com/yapp/orbit/OrbitNavHost.kt
@@ -1,4 +1,4 @@
-package com.yapp.navigator
+package com.yapp.orbit
 
 import android.annotation.SuppressLint
 import androidx.compose.animation.AnimatedVisibility
