+++ build-logic/src/main/java/com/yapp/convention/HiltAndroid.kt
@@ -14,6 +14,7 @@ internal fun Project.configureHiltAndroid() {
     dependencies {
         "implementation"(libs.findLibrary("hilt.android").get())
         "ksp"(libs.findLibrary("hilt.android.compiler").get())
+        "implementation"(libs.findLibrary("hilt-navigation-compose").get())
     }
 }
