+++ build-logic/src/main/java/orbit.android.feature.gradle.kts
@@ -19,7 +19,6 @@ dependencies {
     implementation(project(":core:ui"))
 
     val libs = project.extensions.libs
-    implementation(libs.findLibrary("hilt-navigation-compose").get())
     implementation(libs.findLibrary("compose-navigation").get())
     implementation(libs.findLibrary("lifecycle-viewmodel").get())
     implementation(libs.findLibrary("lifecycle-runtime").get())
