--- feature/navigator/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest>
-</manifest>
