+++ settings.gradle.kts
@@ -35,7 +35,6 @@ include(":domain")
 include(":feature")
 include(":core:ui")
 include(":feature:home")
-include(":feature:navigator")
 include(":feature:onboarding")
 include(":feature:mission")
 include(":feature:fortune")
